factor out the resolve step status logic to a module
